Add features/support/world.js
